+++ backend/core/src/applications/applications.module.ts
@@ -16,11 +16,13 @@ import { ApplicationsService } from "./services/applications.service"
 import { CsvBuilder } from "./services/csv-builder.service"
 import { ApplicationCsvExporterService } from "./services/application-csv-exporter.service"
 import { EmailModule } from "../email/email.module"
+import { ActivityLogModule } from "../activity-log/activity-log.module"
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([Application, Applicant, Address, Listing]),
     AuthModule,
+    ActivityLogModule,
     SharedModule,
     ListingsModule,
     ApplicationFlaggedSetsModule,