+++ sites/public/pages/applications/review/summary.tsx
@@ -6,14 +6,15 @@ import {
   AppearanceStyleType,
   Button,
   FormCard,
-  ProgressNav,
   t,
   Form,
+  ProgressNav,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../../layouts/forms"
 import { useForm } from "react-hook-form"
 import FormSummaryDetails from "../../../src/forms/applications/FormSummaryDetails"
 import { useFormConductor } from "../../../lib/hooks"
+import { OnClientSide } from "@bloom-housing/shared-helpers"
 
 const ApplicationSummary = () => {
   const { conductor, application, listing } = useFormConductor("summary")
@@ -30,9 +31,9 @@ const ApplicationSummary = () => {
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}
           labels={conductor.config.sections.map((label) => t(`t.${label}`))}
+          mounted={OnClientSide()}
         />
       </FormCard>
-
       <FormCard>
         <div className="form-card__lead">
           <h2 className="form-card__title is-borderless">