+++ sites/public/CHANGELOG.md
@@ -3,6 +3,151 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.1-alpha.73](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.72...@bloom-housing/public@3.0.1-alpha.73) (2021-12-15)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.72](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.71...@bloom-housing/public@3.0.1-alpha.72) (2021-12-15)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.71](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.70...@bloom-housing/public@3.0.1-alpha.71) (2021-12-15)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.70](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.69...@bloom-housing/public@3.0.1-alpha.70) (2021-12-14)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.69](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.68...@bloom-housing/public@3.0.1-alpha.69) (2021-12-14)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.68](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.67...@bloom-housing/public@3.0.1-alpha.68) (2021-12-13)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.67](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.65...@bloom-housing/public@3.0.1-alpha.67) (2021-12-13)
+
+
+### Bug Fixes
+
+* versioning issues ([#2311](https://github.com/bloom-housing/bloom/issues/2311)) ([0b1d143](https://github.com/bloom-housing/bloom/commit/0b1d143ab8b17add9d52533560f28d7a1f6dfd3d))
+
+
+
+
+
+## [3.0.1-alpha.65](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.64...@bloom-housing/public@3.0.1-alpha.65) (2021-12-10)
+
+
+### Bug Fixes
+
+* demographics howDidYouHear options were not changing on language change ([#2305](https://github.com/bloom-housing/bloom/issues/2305)) ([a9859d1](https://github.com/bloom-housing/bloom/commit/a9859d1a718d88e0e1ff671a408887691306b4b9))
+
+
+
+
+
+## [3.0.1-alpha.64](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.63...@bloom-housing/public@3.0.1-alpha.64) (2021-12-10)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.63](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.62...@bloom-housing/public@3.0.1-alpha.63) (2021-12-10)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.62](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.61...@bloom-housing/public@3.0.1-alpha.62) (2021-12-09)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.61](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.60...@bloom-housing/public@3.0.1-alpha.61) (2021-12-09)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.60](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.59...@bloom-housing/public@3.0.1-alpha.60) (2021-12-09)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.59](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.58...@bloom-housing/public@3.0.1-alpha.59) (2021-12-09)
+
+
+### Bug Fixes
+
+* prefUnits on autofill summary ([#2294](https://github.com/bloom-housing/bloom/issues/2294)) ([2b370e9](https://github.com/bloom-housing/bloom/commit/2b370e9321ec652ed3d124319e08e01fc765d636))
+
+
+
+
+
+## [3.0.1-alpha.58](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.57...@bloom-housing/public@3.0.1-alpha.58) (2021-12-09)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.57](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.56...@bloom-housing/public@3.0.1-alpha.57) (2021-12-09)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.56](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.55...@bloom-housing/public@3.0.1-alpha.56) (2021-12-08)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
 ## [3.0.1-alpha.55](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.54...@bloom-housing/public@3.0.1-alpha.55) (2021-12-07)
 
 **Note:** Version bump only for package @bloom-housing/public